move some components into files
